pyproject.toml
[project]
name = "promptherd"
version = "0.1.0"
description = "an app to evaluate and share prompts evaluations"
authors = [
    {name = "jonathanpasse@protonmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"

[tool.poetry.dependencies]
python = ">=3.12"
jsonschema = ">=4.25.1,<5.0.0"
litellm = "^1.76.1"
numpy = "^2.3.2"
pyyaml = "^6.0.2"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.4.1,<9.0.0"



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
